chore: restructured database classes
