+++ energetica/utils/climate_helpers.py
@@ -11,8 +11,8 @@ from numpy.typing import ArrayLike
 from energetica.config.climate_events import climate_events
 from energetica.database.active_facility import ActiveFacility
 from energetica.database.climate_event_recovery import ClimateEventRecovery
-from energetica.database.engine_data import calculate_reference_gta, calculate_temperature_deviation
-from energetica.database.map import HexTile
+from energetica.database.emission_data import calculate_reference_gta, calculate_temperature_deviation
+from energetica.database.map.hex_tile import HexTile
 from energetica.enums import FunctionalFacilityType, Renewable
 from energetica.globals import engine
 from energetica.utils.assets import facility_destroyed
