+++ energetica/utils/map_helpers.py
@@ -1,6 +1,6 @@
 """Utility functions relating to the game map."""
 
-from energetica.database.map import HexTile
+from energetica.database.map.hex_tile import HexTile
 from energetica.database.player import Player
 from energetica.game_error import GameError, GameExceptionType
 from energetica.globals import engine
