+++ tests/policy_runner.py
@@ -1,6 +1,6 @@
 """Run a list of policies with one player per policy."""
 
-from energetica.database.map import HexTile
+from energetica.database.map.hex_tile import HexTile
 from energetica.database.player import Player
 from energetica.enums import Renewable
 from energetica.globals import engine
